fix: f64 -> i32

+++ src/eval.rs
@@ -1,6 +1,6 @@
 use crate::parse::{ast::Node, token::Token};
 
-pub fn eval(ast: Node) -> f64 {
+pub fn eval(ast: Node) -> i32 {
     match ast {
         Node::Number(num) => num,
         Node::BinaryOperator(bo) => match bo.op {
